[FLINK-4123] [cassandra] Fix concurrency issue in CassandraTupleWriteAheadSink
The updatesCount variable in the CassandraTupleWriteAheadSink.sendValues did not have
guaranteed visibility. Thus, it was possible that the callback thread would read an
outdated value for updatesCount, resulting in a deadlock. Replacing IntValue updatesCount
with AtomicInteger updatesCount fixes this issue.

Furthermore, the PR hardens the CassandraTupleWriteAheadSinkTest which could have failed
with a NPE if the callback runnable was not set in time.